refactor: 단일 책임 원칙에 따라 API 호출 함수를 분리한다
